+++ pallets/evercity-exchange/src/lib.rs
@@ -114,7 +114,7 @@ pub mod pallet {
 
 	/// Calls:
 	#[pallet::call]
-	impl<T: Config> Pallet<T> where <T as pallet_evercity_assets::pallet::Config>::Balance: From<u128> + Into<u128>  {
+	impl<T: Config> Pallet<T> where <T as pallet_evercity_assets::pallet::Config>::Balance: From<u64> + Into<u64>  {
 
 		/// <pre>
 		/// Method: create_carbon_credit_lot
